Refactor header footer utils
