+++ src/Nodes/Body.ts
@@ -8,7 +8,7 @@ import { Node, NodeViewRendererProps, mergeAttributes } from "@tiptap/core";
 import { BODY_ATTRIBUTES, BODY_NODE_NAME } from "../constants/body";
 import { isBodyNode } from "../utils/nodes/body";
 import { parseHTMLNode } from "../utils/nodes/node";
-import { calculateShorthandMargins, calculateBodyMargins } from "../utils/pageRegion/margins";
+import { calculateShorthandMargins, calculateBodyMargins } from "../utils/nodes/body/attributes/pageMargins";
 import { mm } from "../utils/units";
 import { calculateBodyDimensions } from "../utils/pageRegion/dimensions";
 import { constructChildOnlyClipboardPlugin } from "../utils/clipboard";
