+++ src/Nodes/Page.ts
@@ -8,8 +8,8 @@ import { Node, NodeViewRendererProps, mergeAttributes } from "@tiptap/core";
 import { DEFAULT_PAGE_BORDER_CONFIG } from "../constants/pageBorders";
 import { DEFAULT_PAPER_COLOUR } from "../constants/paperColours";
 import { PAGE_NODE_NAME, DEFAULT_PAGE_GAP, PAGE_ATTRIBUTES } from "../constants/page";
-import { getPaperDimensionsFromPageNode } from "../utils/paperSize";
-import { getPageNodePaperColour } from "../utils/paperColour";
+import { getPaperDimensionsFromPageNode } from "../utils/nodes/page/attributes/paperSize";
+import { getPageNodePaperColour } from "../utils/nodes/page/attributes/paperColour";
 import { isPageNode } from "../utils/nodes/page/page";
 import { mm, px } from "../utils/units";
 import { calculateShorthandPageBorders, getPageNodePageBorders } from "../utils/pageBorders";
