+++ src/index.ts
@@ -32,10 +32,10 @@ import BodyNode from "./Nodes/Body";
 export { isPageNode } from "./utils/nodes/page/page";
 export { getThisPageNodePosition } from "./utils/nodes/page/pagePosition";
 export { getPageNumber } from "./utils/nodes/page/pageNumber";
-export { getPageNumPaperSize, getPageNodePaperSize } from "./utils/paperSize";
-export { getPageNumPaperColour, getPageNodePaperColour } from "./utils/paperColour";
-export { getPageNumPaperOrientation, getPageNodePaperOrientation } from "./utils/paperOrientation";
-export { getPageNumPageMargins } from "./utils/pageRegion/margins";
+export { getPageNumPaperSize, getPageNodePaperSize } from "./utils/nodes/page/attributes/paperSize";
+export { getPageNumPaperColour, getPageNodePaperColour } from "./utils/nodes/page/attributes/paperColour";
+export { getPageNumPaperOrientation, getPageNodePaperOrientation } from "./utils/nodes/page/attributes/paperOrientation";
+export { getPageNumPageMargins } from "./utils/nodes/body/attributes/pageMargins";
 export { getPageNumPageBorders, getPageNodePageBorders } from "./utils/pageBorders";
 export { getBodyNodeMargins } from "./utils/nodes/body";
 export { doesDocHavePageNodes } from "./utils/nodes/page/page";
