+++ src/utils/margins/pageRegion.ts
@@ -7,22 +7,8 @@
 import { Node as PMNode } from "@tiptap/pm/model";
 import { HeaderNodeAttributes, FooterNodeAttributes, HeaderFooter } from "../../types/pageRegions";
 import { Nullable } from "../../types/record";
-import {
-    HEADER_DEFAULT_ATTRIBUTES,
-    FOOTER_DEFAULT_ATTRIBUTES,
-    HEADER_FOOTER_NODE_ATTR_KEYS,
-    HEADER_FOOTER_NODE_NAME,
-} from "../../constants/pageRegions";
+import { HEADER_DEFAULT_ATTRIBUTES, FOOTER_DEFAULT_ATTRIBUTES, HEADER_FOOTER_NODE_ATTR_KEYS } from "../../constants/pageRegions";
 import { XMarginConfig } from "../../types/page";
-
-/**
- * Determines if the given node is a header node.
- * @param node - The node to check.
- * @returns {boolean} True if the node is a header node, false otherwise.
- */
-export const isHeaderFooterNode = (node: PMNode): boolean => {
-    return node.type.name === HEADER_FOOTER_NODE_NAME;
-};
 
 /**
  * Get the type of the header or footer node.
